Start working over this shit, needs to be done, now is shit with adding users
