+++ frontend/src/app/(authorized)/view/aquariums/[id]/page.tsx
@@ -3,19 +3,23 @@
 import MonitorCard, { MonitorCardProps } from "@/components/MonitorCard";
 import { useRouter } from "next/navigation";
 import Image from "next/image";
-import { FirestoreContext } from "@/context/FirebaseProvider";
-import { doc, getDoc } from "firebase/firestore";
-import { useCallback, useContext, useEffect, useState } from "react";
-import { HealthStatus } from "@/enums/HealthStatus.enum";
+import { useEffect } from "react";
+
+import {
+  generateFirstRowData,
+  generateSecondRowData,
+  generateThirdRowData,
+  useAquariumData,
+} from "./data.logic";
 import { AquariumDataProps } from "../../page";
 
 interface AquariumAquaViewPageProps {
   params: { id: string };
 }
 
-type CardData = {
+export type CardData = {
   column: string;
-  value: string;
+  value: string | undefined;
   iconUrl: string;
 };
 
@@ -23,67 +27,17 @@ export default function AquariumAquaViewPage({
   params,
 }: AquariumAquaViewPageProps) {
   const router = useRouter();
-  const firestore = useContext(FirestoreContext);
 
-  const [aquariumData, setAquariumData] = useState<
-    AquariumDataProps | Record<string, any>
-  >({});
-
-  const getAquariumData = useCallback(async () => {
-    if (!firestore) return;
-
-    const aquariumId = params.id;
-    const aquariumRef = doc(firestore, "aquariums", aquariumId);
-
-    const aquariumSnapshot = await getDoc(aquariumRef);
-
-    const aquariumData = aquariumSnapshot.data();
-
-    const userId = aquariumData?.user_id;
-
-    const usersRef = doc(firestore, "users", userId);
-
-    const userSnapshot = await getDoc(usersRef);
-
-    const user = userSnapshot.data();
-
-    const aquariumSize =
-      (aquariumData?.width / 100) *
-        (aquariumData?.height / 100) *
-        (aquariumData?.length / 100) +
-      "m^3";
-
-    //TODO: how is healthStatus prepared?
-    const healthStatus = HealthStatus.GOOD;
-    const mappedAquariumData = {
-      id: aquariumId,
-      name: user?.username || "",
-      avatar: "",
-      email: user?.email || "",
-      aquariumTitle: aquariumData?.name,
-      healthStatus,
-      aquariumSize,
-      isLiked: user?.fav_aquariums?.includes(aquariumId) || false,
-      aquariumData: {
-        fishes: aquariumData?.fishes,
-        pump: aquariumData?.pump,
-        heater: aquariumData?.heater,
-        light: aquariumData?.light,
-        plants: aquariumData?.plants,
-        decors: aquariumData?.decors,
-        terrains: aquariumData?.terrains,
-      },
-    };
-
-    setAquariumData(mappedAquariumData);
-
-    // eslint-disable-next-line react-hooks/exhaustive-deps
-  }, []);
+  const { aquariumData, getAquariumData } = useAquariumData(params);
 
   useEffect(() => {
     getAquariumData();
     // eslint-disable-next-line react-hooks/exhaustive-deps
   }, []);
+
+  const firstRowData = generateFirstRowData(aquariumData);
+  const secondRowData = generateSecondRowData(aquariumData);
+  const thirdRowData = generateThirdRowData(aquariumData);
 
   // TODO: get this data from backend
   const data = [
@@ -148,80 +102,6 @@ export default function AquariumAquaViewPage({
       </button>
     </>
   );
-
-  const firstRowData = [
-    {
-      column: "Aquarium title",
-      value: aquariumData?.aquariumTitle,
-      iconUrl: "/title.svg",
-    },
-    {
-      column: "Aquarium size",
-      value: aquariumData?.aquariumSize,
-      iconUrl: "/size.svg",
-    },
-    {
-      column: "Health status",
-      value: aquariumData?.healthStatus,
-      iconUrl: "/health-status.svg",
-    },
-  ];
-  const secondRowData = [
-    {
-      column: "Pump",
-      value: aquariumData?.aquariumData?.pump?.name,
-      iconUrl: "/pump.svg",
-    },
-    {
-      column: "Heater",
-      value: aquariumData?.aquariumData?.heater?.name,
-      iconUrl: "/heater.svg",
-    },
-    {
-      column: "Light",
-      value: aquariumData?.aquariumData?.light?.name,
-      iconUrl: "/light.svg",
-    },
-  ];
-
-  const thirdRowData = [
-    {
-      column: "Plants",
-      value:
-        aquariumData?.aquariumData?.plants &&
-        aquariumData?.aquariumData?.plants
-          .map((plants: any) => plants.name)
-          .join(", "),
-      iconUrl: "/plants.svg",
-    },
-    {
-      column: "Decors",
-      value:
-        aquariumData?.aquariumData?.decors &&
-        aquariumData?.aquariumData?.decors
-          .map((decor: any) => decor.name)
-          .join(", "),
-      iconUrl: "/decors.svg",
-    },
-    {
-      column: "Terrains",
-      value:
-        aquariumData?.aquariumData?.terrains &&
-        aquariumData?.aquariumData?.terrains
-          .map((terrain: any) => terrain.name)
-          .join(", "),
-      iconUrl: "/terrain.svg",
-    },
-    {
-      column: "Fishes",
-      value:
-        aquariumData?.aquariumData?.fishes &&
-        aquariumData?.aquariumData?.fishes
-          .map((fish: any) => fish.name)
-          .join(", "),
-      iconUrl: "/fishes.svg",
-    },
-  ];
 
   const Cards = ({ data }: { data: CardData[] }) => {
     return (
